Fixed Adafruit_GFX font structure size limitation

+++ src/bb_spi_lcd.h
@@ -86,11 +86,11 @@ typedef void (*RESETCALLBACK)(void);
 #define _GFXFONT_H_
 typedef struct {
   uint16_t bitmapOffset; ///< Pointer into GFXfont->bitmap
-  uint8_t width;         ///< Bitmap dimensions in pixels
-  uint8_t height;        ///< Bitmap dimensions in pixels
-  uint8_t xAdvance;      ///< Distance to advance cursor (x axis)
-  int8_t xOffset;        ///< X dist from cursor pos to UL corner
-  int8_t yOffset;        ///< Y dist from cursor pos to UL corner
+  uint16_t width;         ///< Bitmap dimensions in pixels
+  uint16_t height;        ///< Bitmap dimensions in pixels
+  uint16_t xAdvance;      ///< Distance to advance cursor (x axis)
+  int16_t xOffset;        ///< X dist from cursor pos to UL corner
+  int16_t yOffset;        ///< Y dist from cursor pos to UL corner
 } GFXglyph;
 
 /// Data stored for FONT AS A WHOLE
@@ -99,7 +99,7 @@ typedef struct {
   GFXglyph *glyph;  ///< Glyph array
   uint8_t first;    ///< ASCII extents (first char)
   uint8_t last;     ///< ASCII extents (last char)
-  uint8_t yAdvance; ///< Newline distance (y axis)
+  uint16_t yAdvance; ///< Newline distance (y axis)
 } GFXfont;
 #endif // _ADAFRUIT_GFX_H
